Refactor: CommentLIst Component 위치 변경
