+++ web/pages/folderDetail/components/index.ts
@@ -1,3 +1,2 @@
-export { default as CommentList } from './CommentList';
 export { default as ContentSection } from './ContentSection';
 export { default as CommentSection } from './CommentSection';
